+++ control-core/src/test/kotlin/at/florianschuster/control/StubTest.kt
@@ -22,7 +22,7 @@ internal class StubTest {
 
     @Test
     fun `custom controller implementation cannot be stubbed`() {
-        val sut = object : Controller<Int, Int, Int> {
+        val sut = object : Controller<Int, Int> {
             override fun dispatch(action: Int) = Unit
             override val currentState: Int get() = error("")
             override val state: Flow<Int> get() = error("")
@@ -103,7 +103,7 @@ internal class StubTest {
     private fun CoroutineScope.createStringController() =
         ControllerImplementation<List<String>, List<String>, List<String>>(
             scope = this,
-            dispatcher = scopeDispatcher,
+            dispatcher = defaultScopeDispatcher(),
             controllerStart = ControllerStart.Lazy,
             initialState = initialState,
             mutator = { flowOf(it) },